backend/requirements.txt

flask
flask-cors
langchain
openai
python-dotenv